renamed communication folder to event folder and things are working
