+++ src/engine/core/events/Events.ts
@@ -1,0 +1,3 @@
+export namespace Events {
+    export const SCENE_CHANGE: string = "SCENE_CHANGE";
+}
